+++ examples/resources/nxos_queuing_qos_policy_map_match_class_map/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_queuing_qos_policy_map_match_class_map.example "sys/ipqos/queuing/p/name-[PM1]/cmap-[c-out-q1]"